src/br/ufsc/ine/leb/roza/SimilarityReport.java
package br.ufsc.ine.leb.roza;

import java.util.ArrayList;
import java.util.Collections;
import java.util.List;

public class SimilarityReport {

	private List<SimilarityAssessment> assessments;

	public SimilarityReport(List<SimilarityAssessment> assessments) {
		this.assessments = new ArrayList<>(assessments);
	}

	public List<SimilarityAssessment> getAssessments() {
		return Collections.unmodifiableList(assessments);
	}

	public void removeReflexiveAssessments() {
		assessments.removeIf((assessment) -> {
			return assessment.getSource().equals(assessment.getTarget());
		});
	}

}
